Fix build errors and compiler warnings across multiple files
- CPose3D_unittest: fix Eigen jacobiSvd API for newer Eigen (use JacobiSVD constructor)
- CIbeoLuxETH: use float casts matching convertToCartesian() signature
- ReactiveNav3D_demo: fix sign()*remainder() double/float mixing
- RawLogViewer: fix size_t/float/TCoords conversion warnings
- robot-map-gui: fix int-to-float conversions in GL widget code
- simul-landmarks: fix size_t loop variable for matrix indexing
- vector_with_small_size_optimization: suppress GCC false-positive array-bounds warning
- hwdrivers_kinect example: use float literal for setScale()

+++ apps/mrpt_apps_gui/RawLogViewer/main_timeline.cpp
@@ -381,10 +381,10 @@ void xRawLogViewerFrame::bottomTimeLineUpdateCursorFromTreeScrollPos()
 
   // percent of view:
   double pc0 = 0, pc1 = 0;
-  if (const double nItems = m_treeView->getTotalTreeNodes(); nItems > 2)
+  if (const double nItems = static_cast<double>(m_treeView->getTotalTreeNodes()); nItems > 2)
   {
-    pc0 = m_treeView->m_firstVisibleItem / static_cast<double>(nItems - 1);
-    pc1 = m_treeView->m_lastVisibleItem / static_cast<double>(nItems - 1);
+    pc0 = static_cast<double>(m_treeView->m_firstVisibleItem) / (nItems - 1);
+    pc1 = static_cast<double>(m_treeView->m_lastVisibleItem) / (nItems - 1);
   }
 
   // visible page shaded area:
